Many more fixes
