+++ serve.ts
@@ -90,10 +90,43 @@ const server = Bun.serve({
 		}
 
 		if (url.pathname === "/upload") {
-			console.info("Upload something");
-			console.info(req);
+			// Something file related - SD storage
+			const searchParams = url.searchParams;
+			const path = searchParams.get("path");
+
+			if (req.method === "GET") {
+				const dummyFileList = {
+					"files": [
+						{ "name": "CoasterHolesPartBoard-gcc.nc", "shortname": "CoasterHolesPartBoard-gcc.nc", "size": "3515", "datetime": "" },
+						{ "name": "CoasterHolesFullBoard.nc", "shortname": "CoasterHolesFullBoard.nc", "size": "52622", "datetime": "" }
+					], "path": "", "total": "119.00 MB", "used": "57.00 KB", "occupation": "0", "status": "Ok"
+				};
+				return new Response(JSON.stringify(dummyFileList), { status: 200, headers: { contentType: "application/json" } });
+			}
+			console.log(req);
 			return new Response("", { status: 200 });
 		}
+
+		if (url.pathname === "/files") {
+			// Something file related - regular storage
+			const searchParams = url.searchParams;
+			const action = searchParams.get("action");
+			const path = searchParams.get("path");
+
+			if (action === "list") {
+				const dummyFileList = {
+					"files": [
+						{ "name": "config-bak.yaml", "shortname": "config-bak.yaml", "size": "3585", "datetime": "" },
+						{ "name": "favicon.ico", "shortname": "favicon.ico", "size": "1150", "datetime": "" },
+						{ "name": "index.html.gz", "shortname": "index.html.gz", "size": "113340", "datetime": "" },
+						{ "name": "maslow.yaml", "shortname": "maslow.yaml", "size": "3546", "datetime": "" }
+					],
+					"path": "", "total": "192.00 KB", "used": "132.00 KB", "occupation": "68", "status": "Ok"
+				};
+				return new Response(JSON.stringify(dummyFileList), { status: 200, headers: { contentType: "application/json" } });
+			}
+		}
+
 
 		const checkFileBase = `./www${url.pathname}`;
