+++ www/js/common.js
@@ -174,7 +174,8 @@ import {
 	tabletShowMessage,
 	tabletUpdateModal,
 } from "./tablet.js";
-import { arrayToXYZ, displayer, refreshGcode } from "./toolpath-displayer.js";
+import { drawTPBtns } from "./tabletControls.js";
+import { arrayToXYZ, refreshGcode, tpDisplayer, tpInit } from "./toolpath-displayer.js";
 import { translate_text, trans_text_item } from "./translate.js";
 import { UIdisableddlg } from "./UIdisableddlg.js";
 import { updatedlg } from "./updatedlg.js";
@@ -495,8 +496,10 @@ export {
 	tabletGrblState,
 	tabletShowMessage,
 	tabletUpdateModal,
+	// from "./tabletControls.js";
+	drawTPBtns,
 	// from "./toolpath-displayer.js";
-	arrayToXYZ, displayer, refreshGcode,
+	arrayToXYZ, refreshGcode, tpDisplayer, tpInit,
 	// from translate.js
 	translate_text, trans_text_item,
 	// from UIdisableddlg.js
